ignore debug and release; changed type errors too

+++ src/dealer.cpp
@@ -38,7 +38,7 @@ int Dealer::get_hand_size() {
 
 void Dealer::print_hand(std::ostream& output) {
     output << "Dealers hand: [";
-    for (int i = 0; i < this->hand.size(); i++) {
+    for (size_t i = 0; i < this->hand.size(); i++) {
         if (i == this->hand.size()-1) {
             output << this->hand[i];
             break;
